Create Header component
This is essentially a static component that replicates the header from
the mockup in react, using some slight variations on the existing css to
simplify the layouts. Since we're going to display this on fixed-size
monitors (1920x1080), I've opted to use hard-coded height/widths in a
number of places and left out some of the @media queries that were used
initially.