Implement showing/hiding of keyboard
We're using a listener on the body to detect taps/click "anywhere", and
translating the position of those taps into a valid position where we
can place our keyboard. Once the keyboard is rendered, we add an
additional listener on the background to dismiss the keyboard, which
prevents it from hiding when tapping on the keys/buttons themselves.

I'm not 100% on how this will affect the pagination buttons (if at all),
so we may need to revisit this in the future to make sure we can prevent
the event from propagating up or down in the stack.