Reload the app after inactive interval
Creates a useEffect hook that sets an inactivity timeout, which is
then unset in the cleanup function -- the end result being that any
change in the state values causes the timeout to restart.

I also had to move the keyboard visibility into the App component. I was
originally trying to keep all of the keyboard functionality in the
OnScreenKeyboard component, but I found that I needed to move the
isVisible state up into the App in order to have the inactivity timer
restart properly when the keyboard opens -- Otherwise some unlucky user
could run into the case of the window reloading in the time between
tapping the app to bring up the keyboard and actually entering a search
query.