Implement basic Algolia search
This is a very barebones Algolia implementation. It just accepts a
searchQuery value as prop and dumps out stringified results JSON into
the DOM. Future tickets will address the specifics for styling,
pagination, etc.

Running this code does require creating a file at
`src/const/algolia.ts` that populates the appid, api key, and
index name (based on the index file provided). These are not, strictly
speaking, secret values since they'll be exposed on the client anyway,
but this does allow us to modify the algolia details at compile time if
we need to change the target.

For the API key, we should generate a [Secured API Key][1] from the Algolia
dashboard and use that, instead of the primary API key.

One edge case note: The search input above the keyboard is technically a
regular input, so you can focus it and type in text with your physical
keyboard, if desired. However, that doesn't actually update the
searchInput state value, so attempting to search that text will return
all the results. That shouldn't be a problem in production, when there
is no physical keyboard, but it did trip me up a few times while
testing. The other option would be to set the input as read only, but
that also removes the ability to tap the input text and move the cursor.

[1]: https://www.algolia.com/doc/guides/security/api-keys/#secured-api-keys