Initial CI/CD workflows
This sets up a basic CI/CD workflow that runs all the tests, then builds
the app bundle and uploads the complete bundle as a build artifact. This
can be downloaded from the workflow page, thoguht it will only be
retained for 90 days. I'm going to continue exploring other ways to
attach the bundle to a release so it will persist longer.