Get correct release ID
The object returned from the `github.*` calls are http response style
objects, so the ID, url, etc. are all under `reponse.data`, not directly
under `response`. This updates the code to destructure those values
correctly.

This also adds some normal info output that provides the release URL and
the URL of the uploaded code package.