Split enter and caps handlers
Both of these handlers were being run on the keyPressed event, which
was leading to the problem with phantom pointerUp events re-opening the
modal. I then moved them to the keyReleased event, but that was causing
tests to fail. Instead, we'll just split them into separate methods and
run them on the appropriate events.