Use user-event for click actions
The @testing-library/user-event library has a more robust way of
handling click/touch/pointer events in testing that triggers all stages
of the event, which more closely resembles how a user would use the
application. It's also clearer because we don't have to call
fireEvent.pointerDown for some events and fireEvent.click for others.

However, using userEvent as a direct replacement triggers a lot of
warnings in the test output about needing to wrap them in act().
According to a blog post from the author[1], this is caused by side
effect triggered by the actions leading to "unexpected" changes in the
output. I think a lot of those side effects are generated from within
the react-simple-keyboard library, so they're not really things that we
can change. Given that, we can just wrap everything in act() like it
says, which quiets the errors (and seems to make the rendering events
more consistent).

[1]: https://kentcdodds.com/blog/fix-the-not-wrapped-in-act-warning