Move SearchCount out of SearchPagination
This is mainly to make it easier to unit test the individual
functionality of the Pagination Component without needing an algolia
instance to wrap around the SearchCount. I moved the div that styles the
wrapper around the count into the Pagination component as its css rules
are more closely connected to the css for the Pagination.