Calculate next/prev page in callbacks
It's possible that the values for nextPage and prevPage were not being
updated as soon as the props changes, and that the old values were being
passed to `refine` instead. This changes the structure so that the
argument to `refine` is being calculated in the callback handler
directly, instead of being stored in a variable.