Add concurrency group
The [`jobs.<job_id>.concurrency`][1] option lets you define a
"concurrency" group, for which there will only ever be one job running
in that that same group. When a second job tries to run, it will be
marked as "pending" and wait for the first to finish.

It looks like the situation is a little different for a third job, as
any new jobs initiated in that concurrency group will cancel any
existing pending jobs. So for any n number of jobs initiated past the
first, only that first and the last will run. This would break our
record of minor version bumps for each PR merge, but I think that's a
rare enough situation that it shouldn't be a problem.

[1]: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idconcurrency