Always build and deploy from main
The previous configuration built from the specific commit that triggered
the workflow run, but if that's not necessarily the most recent commit
to the branch then we won't be able to commit the version bump back in
the later steps.

Instead, we'll always check out the current HEAD of main and work from
there.