Check for [release] tag in commit
This adds a way to opt out of the version bump stage of the job for
any commit that starts with [release]. Github will already ignore
commits from GitHub Actions, so this isn't really changing any behavior
for our automatic bumps (though I've added it to the commit message for
consistency). Instead, the purpose is to let us manually push a new
version wihouth immediately increasing the version again.

For example, after the PR containing this commit is merged I plan to
push a new commit directly to main that will bump the version number to
`v1.0.0` since we'll be ready for our "formal" release. Without this
change, that would trigger a new workflow run and immediately bump the
version to `v1.0.1`. This is not the end of the world, but it does give
us a little more control over how and when the workflow runs.

The rest of the workflow would run as normal, so our `v1.0.0` release
would still be packaged and published. I did consider adding some
additional code to the github-script in the "Publish Bundle" section to
check for an existing release with the same version tag before creating
a new one, but I don't think that's strictly necessary at the moment --
we'll see if that becomes necessary in the future.